+++ src/pook/__init__.py
@@ -44,4 +44,4 @@ __author__ = "Tomas Aparicio"
 __license__ = "MIT"
 
 # Current version
-__version__ = "2.1.1"
+__version__ = "2.1.2"